requirements.txt
aiohttp>=3.9.0
json5>=0.9.14
Pillow>=10.0.0
